requirements.txt
yfinance
Pystan==2.19.1.1
fbprophet
streamlit
plotly
hvplot
pandas.core.frame